pyproject.toml
[tool.poetry]
name = "image-classifier"
version = "0.1.0"
description = ""
authors = ["Christos Natsis <christos_na@hotmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
tensorflow = "^2.14.0"
numpy = "^1.26.2"
opencv-python = "^4.8.1.78"
tensorflow-macos = { version = "^2.15.0", platform = "darwin", markers = "platform_machine=='arm64'" }
tensorflow-intel = { version = "^2.14.0", platform = "win32" }
tensorflow-cpu = [
    { version = "^2.14.0", platform = "linux", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" },
    { version = "^2.14.0", platform = "darwin", markers = "platform_machine!='arm64' and platform_machine!='aarch64'" },]
tensorflow-cpu-aws = { version = "^2.14.0", platform = "linux", markers = "platform_machine=='arm64' or platform_machine=='aarch64'" }
# https://github.com/tensorflow/tensorflow/blob/adb39b04e9cb116df4659a7e2de9eea27e62f25c/tensorflow/tools/pip_package/setup.py#L107-L108
# https://github.com/python-poetry/poetry/issues/8271#issuecomment-1697740447
tensorflow-io-gcs-filesystem = [
    { version = ">= 0.23.1", markers = "platform_machine!='arm64' or platform_system!='Darwin'" },
    { version = "< 0.32.0", markers = "platform_system == 'Windows'" }
]
pillow = "^10.1.0"
scipy = "^1.11.3"
tf2onnx = "^1.15.1"
onnx = "1.14.1"
onnxruntime = "^1.16.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
